+++ HospitalCase.WebAPI/Controllers/MedicalRecordsController.cs
@@ -1,7 +1,10 @@
 ﻿using HospitalCase.WebAPI.Interfaces;
 using HospitalCase.WebAPI.Models;
+using HospitalCase.WebAPI.Repositories;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -23,40 +26,99 @@ namespace HospitalCase.WebAPI.Controllers
 
         // GET: api/<MedicalRecordsController>
         [HttpGet]
-        public async Task<IEnumerable<MedicalRecord>> Get()
+        public async Task<IActionResult> Get()
         {
-            var result = await _medicalRecordsRepository.GetAllAsync();
-            return result;
+            try
+            {
+                var result = await _medicalRecordsRepository.GetAllAsync();
+                return Ok(result);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
         }
 
         // GET api/<MedicalRecordsController>/5
         [HttpGet("{id}")]
-        public async Task<MedicalRecord> Get(int id)
+        public async Task<IActionResult> Get(int id)
         {
-            var result = await _medicalRecordsRepository.GetByIdAsync(id);
+            if (id <= 0) return BadRequest();
 
-            return result;
+            try
+            {
+                var foundEntry = await _medicalRecordsRepository.GetByIdAsync(id);
+
+                if (foundEntry == null) return NotFound();
+
+                return Ok(foundEntry);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
         }
 
         // POST api/<MedicalRecordsController>
         [HttpPost]
-        public async void Post([FromBody] MedicalRecord medicalRecord)
+        public async Task<IActionResult> Post([FromBody] MedicalRecord medicalRecord)
         {
-            await _medicalRecordsRepository.CreateAsync(medicalRecord);
+            try
+            {
+                await _medicalRecordsRepository.CreateAsync(medicalRecord);
+                return CreatedAtAction(nameof(Get), new { id = medicalRecord.Id }, medicalRecord);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
         }
 
         // PUT api/<MedicalRecordsController>/5
         [HttpPut("{id}")]
-        public async void Put(int id, [FromBody] MedicalRecord medicalRecord)
+        public async Task<IActionResult> Put(int id, [FromBody] MedicalRecord medicalRecord)
         {
-            await _medicalRecordsRepository.UpdateAsync(id, medicalRecord);
+            if (id != medicalRecord.Id) return BadRequest();
+
+            try
+            {
+                var foundEntry = await _medicalRecordsRepository.GetByIdAsync(id);
+
+                if (foundEntry == null) return NotFound();
+
+                await _medicalRecordsRepository.UpdateAsync(id, medicalRecord);
+
+                return NoContent();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
         }
 
         // DELETE api/<MedicalRecordsController>/5
         [HttpDelete("{id}")]
-        public async void Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            await _medicalRecordsRepository.DeleteAsync(id);
+            try
+            {
+                var foundEntry = await _medicalRecordsRepository.GetByIdAsync(id);
+
+                if (foundEntry == null) return NotFound();
+
+                await _medicalRecordsRepository.DeleteAsync(id);
+
+                return NoContent();
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            }
         }
     }
 }
